Issue #1: Client: Added tailwind (with demo markups), added router & pages, removed unused dependencies.
