+++ @shared-types/app-info.json
@@ -1,1 +1,1 @@
-{ "versionInfo": "v.0.0.2 / 2025.06.08 05:01:29 +0300" }
+{ "versionInfo": "v.0.0.2 / 2025.06.08 06:04:58 +0300" }
