+++ client/package.json
@@ -2,7 +2,7 @@
   "name": "client",
   "private": true,
   "version": "0.0.2",
-  "timestamp": "2025.06.08 05:01:29 +0300",
+  "timestamp": "2025.06.08 06:04:58 +0300",
   "changed": "2025.06.07, 19:45",
   "type": "module",
   "main": "src/main.tsx",
@@ -11,11 +11,12 @@
   },
   "dependencies": {
     "react": "^19.1.0",
-    "react-dom": "^19.1.0"
+    "react-dom": "^19.1.0",
+    "react-router-dom": "^7.6.2"
   },
   "devDependencies": {
     "@eslint/js": "^9.25.0",
-    "@tailwindcss/postcss": "^4.1.8",
+    "@tailwindcss/vite": "^4.1.8",
     "@types/node": "^22.15.30",
     "@types/react": "^19.1.2",
     "@types/react-dom": "^19.1.2",
@@ -27,6 +28,7 @@
     "eslint-plugin-react-refresh": "^0.4.19",
     "globals": "^16.0.0",
     "tailwindcss": "^4.1.8",
+    "tinycolor2": "^1.6.0",
     "typescript": "~5.8.3",
     "typescript-eslint": "^8.30.1",
     "vite": "^6.3.5",
