+++ client/src/index.css
@@ -1,3 +1,5 @@
+@import 'tailwindcss';
+
 :root {
   font-family: system-ui, Avenir, Helvetica, Arial, sans-serif;
   line-height: 1.5;
@@ -25,15 +27,18 @@ a:hover {
 body {
   margin: 0;
   display: flex;
-  place-items: center;
+  align-items: stretch;
+  /* place-items: center; */
   min-width: 320px;
   min-height: 100vh;
 }
 
+/*
 h1 {
   font-size: 3.2em;
   line-height: 1.1;
 }
+*/
 
 button {
   border-radius: 8px;
