+++ client/src/pages/Home.tsx
@@ -1,0 +1,13 @@
+// import React from 'react';
+
+export function Home() {
+  return (
+    <div>
+      <h1 className="text-3xl font-bold underline">
+        {/* Tailwind test */}
+        Hello world!
+      </h1>
+      <p>Welcome to the home page!</p>
+    </div>
+  );
+}
