+++ package.json
@@ -3,7 +3,7 @@
   "description": "Dynamic list implemented via ExpressJS and React with a Vercel-based hosting",
   "main": "src/api/index.ts",
   "version": "0.0.2",
-  "timestamp": "2025.06.08 05:01:29 +0300",
+  "timestamp": "2025.06.08 06:04:58 +0300",
   "changed": "2025.06.07, 19:45",
   "keywords": [],
   "author": "lilliputten@gmail.com",
@@ -11,27 +11,18 @@
   "engines": {
     "node": "20.x"
   },
-  "dependencies": {
+  "devDependencies": {
     "@babel/eslint-parser": "^7.25.9",
-    "@commitlint/cli": "^19.6.0",
-    "@commitlint/config-conventional": "^19.6.0",
     "@ianvs/prettier-plugin-sort-imports": "^4.4.0",
-    "@next/eslint-plugin-next": "^14.2.18",
-    "@prisma/client": "^6.0.0",
-    "@tailwindcss/line-clamp": "^0.4.4",
-    "@tailwindcss/typography": "^0.5.15",
-    "@types/jest": "^29.5.14",
-    "@types/ms": "^0.7.34",
     "@types/node": "^20.17.9",
     "@types/tinycolor2": "^1.4.6",
     "@typescript-eslint/eslint-plugin": "^8.16.0",
     "@typescript-eslint/parser": "^8.16.0",
+    "concurrently": "^9.1.2",
     "cross-env": "^7.0.3",
     "editorconfig-checker": "^5.1.9",
     "eslint": "^8.57.1",
-    "eslint-config-next": "15.0.3",
     "eslint-config-prettier": "^9.1.0",
-    "eslint-plugin-jest": "^28.9.0",
     "eslint-plugin-json": "^4.0.1",
     "eslint-plugin-json-files": "^4.4.2",
     "eslint-plugin-jsonc": "^2.18.2",
@@ -39,28 +30,16 @@
     "eslint-plugin-react": "^7.37.2",
     "eslint-plugin-tailwindcss": "^3.17.5",
     "eslint-plugin-yaml": "^1.0.3",
-    "express": "^5.1.0",
     "npm-run-all": "^4.1.5",
     "prettier": "^3.4.1",
     "prettier-plugin-tailwindcss": "^0.6.9",
     "rimraf": "^6.0.1",
-    "sass": "^1.81.0",
     "stylelint": "^16.11.0",
     "stylelint-config-recommended": "^14.0.1",
     "stylelint-config-recommended-scss": "^14.1.0",
     "stylelint-config-tailwindcss": "^0.0.7",
     "stylelint-order": "^6.0.4",
-    "stylelint-scss": "^6.10.0",
-    "tailwind-merge": "^2.5.5",
-    "tailwindcss": "^3.4.15",
-    "tailwindcss-animate": "^1.0.7",
-    "tinycolor2": "^1.6.0",
-    "typescript": "^5.7.2"
-  },
-  "devDependencies": {
-    "@types/express": "^5.0.3",
-    "@types/node": "^22.15.30",
-    "concurrently": "^9.1.2"
+    "stylelint-scss": "^6.10.0"
   },
   "scripts": {
     "postinstall-TODO": "pnpm run prisma-generate && pnpm run prisma-db-push",
