+++ server/package.json
@@ -1,7 +1,7 @@
 {
   "name": "server",
   "version": "0.0.2",
-  "timestamp": "2025.06.08 05:01:29 +0300",
+  "timestamp": "2025.06.08 06:04:58 +0300",
   "changed": "2025.06.08, 05:31",
   "description": "",
   "main": "src/index.ts",
